Add separate manager modules
